deployment/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# SNMP
pysnmp==4.4.12
pysnmp-mibs==0.1.6

# Configuration and environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# WebSocket support
websockets==12.0

# Background tasks and scheduling
celery==5.3.4
redis==5.0.1

# File handling and utilities
aiofiles==23.2.1
openpyxl==3.1.2
reportlab==4.0.7

# HTTP client
httpx==0.25.2
requests==2.31.0

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# CORS support
fastapi-cors==0.0.6

# Date and time utilities
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# Validation and parsing
email-validator==2.1.0